store: Move create_relational_schema into Mapping
